requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
pinecone-client
faiss-cpu
langchain_google_genai